+++ changelog.md
@@ -1,2 +1,8 @@
-- Supports Famers Delight Refabricated
-- Updated a mod that adds carpets
+#### UPDATED
+- Fix render type of RFM Mailboxes to use cutout
+
+#### NEW
+- Mighty Mail (Forge & Fabric)
+- Dawn of Time (Forge)
+- Variant Crafting Tables (Forge)
+- More Crafting Tables (Forge & Fabric)